[SMTChecker] Add SortProvider
